requirements.txt
geopy==1.10.0
Pillow==2.8.1